#2489 - create new `T2SwapCellKiller` class and use it in the vertex code, so that cells removed via a T2 swap are correctly counted in `RemoveDeadCells()`
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@21175 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ cell_based/src/population/VertexBasedCellPopulation.cpp
@@ -223,9 +223,20 @@ unsigned VertexBasedCellPopulation<DIM>::RemoveDeadCells()
     {
         if ((*it)->IsDead())
         {
-            // Remove the element from the mesh
+            // Count the cell as dead
             num_removed++;
-            mpMutableVertexMesh->DeleteElementPriorToReMesh(this->GetLocationIndexUsingCell((*it)));
+
+            // Remove the element from the mesh if it is not deleted yet
+            ///\todo (#2489) this should cause an error - we should fix this!
+            if (!(this->GetElement(this->GetLocationIndexUsingCell((*it)))->IsDeleted()))
+            {
+                // This warning relies on the fact that there is only one other possibility for
+                // vertex elements to be marked as deleted: a T2 swap
+                WARNING("A Cell is removed without performing a T2 swap. This leaves a void in the mesh.");
+                mpMutableVertexMesh->DeleteElementPriorToReMesh(this->GetLocationIndexUsingCell((*it)));
+            }
+
+            // Delete the cell
             it = this->mCells.erase(it);
         }
         else
